+++ frontend/src/scenes/dashboard/dashboardLogic.test.ts
@@ -1,9 +1,9 @@
 import { BuiltLogic } from 'kea'
 import { defaultAPIMocks, mockAPI } from 'lib/api.mock'
-import { expectLogic, initKeaTestLogic } from '~/test/kea-test-utils'
+import { expectLogic, truth } from 'kea-test-utils'
+import { initKeaTestLogic } from '~/test/init'
 import { dashboardLogic } from 'scenes/dashboard/dashboardLogic'
 import dashboardJson from './__mocks__/dashboard.json'
-import { truth } from '~/test/kea-test-utils/jest'
 import { dashboardLogicType } from 'scenes/dashboard/dashboardLogicType'
 import { dashboardsModel } from '~/models/dashboardsModel'
 import { dashboardItemsModel } from '~/models/dashboardItemsModel'
