+++ frontend/src/scenes/insights/InsightMetadata/insightMetadataLogic.test.ts
@@ -1,9 +1,9 @@
 import { BuiltLogic } from 'kea'
 import { insightMetadataLogicType } from 'scenes/insights/InsightMetadata/insightMetadataLogicType'
 import { insightMetadataLogic, InsightMetadataLogicProps } from 'scenes/insights/InsightMetadata/insightMetadataLogic'
-import { expectLogic, initKeaTestLogic } from '~/test/kea-test-utils'
+import { expectLogic, truth } from 'kea-test-utils'
+import { initKeaTestLogic } from '~/test/init'
 import { insightLogic } from 'scenes/insights/insightLogic'
-import { truth } from '~/test/kea-test-utils/jest'
 import { defaultAPIMocks, mockAPI } from 'lib/api.mock'
 import { userLogic } from 'scenes/userLogic'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
