+++ frontend/src/scenes/persons/personHeaderLogic.test.ts
@@ -1,6 +1,7 @@
 import { BuiltLogic } from 'kea'
 
-import { expectLogic, initKeaTestLogic } from '~/test/kea-test-utils'
+import { expectLogic } from 'kea-test-utils'
+import { initKeaTestLogic } from '~/test/init'
 import { PersonType } from '~/types'
 import { uuid } from 'lib/utils'
 import { personHeaderLogic } from 'scenes/persons/personHeaderLogic'
