+++ frontend/src/scenes/sessionRecordings/sessionRecordingsTableLogic.test.ts
@@ -7,7 +7,8 @@ import {
 import { sessionRecordingsTableLogicType } from './sessionRecordingsTableLogicType'
 import { BuiltLogic } from 'kea'
 import { mockAPI, defaultAPIMocks } from 'lib/api.mock'
-import { expectLogic, initKeaTestLogic } from '~/test/kea-test-utils'
+import { expectLogic } from 'kea-test-utils'
+import { initKeaTestLogic } from '~/test/init'
 import { router } from 'kea-router'
 import { PropertyOperator } from '~/types'
 import { RecordingWatchedSource } from 'lib/utils/eventUsageLogic'
