--- frontend/src/test/kea-test-utils/functions/toNotHaveDispatchedActions.ts
@@ -1,13 +1,0 @@
-import { ActionToDispatch, ExpectFunction, testUtilsContext } from '~/test/kea-test-utils'
-import { tryToSearchActions } from '~/test/kea-test-utils/functions/toDispatchActionsInAnyOrder'
-
-export const toNotHaveDispatchedActions: ExpectFunction<ActionToDispatch[]> = {
-    common(logic, actions) {
-        const { notFound } = tryToSearchActions(logic, actions)
-
-        if (notFound.length !== actions.length) {
-            throw new Error(`Found actions when we shouldn't have!`)
-        }
-        testUtilsContext().ranActions = true
-    },
-}
