--- frontend/src/test/kea-test-utils/index.ts
@@ -1,4 +1,0 @@
-export * from './expect'
-export * from './init'
-export * from './plugin'
-export * from './types'
