+++ package.json
@@ -83,7 +83,7 @@
         "kea-loaders": "^0.5.1",
         "kea-localstorage": "^1.1.1",
         "kea-router": "^1.0.2",
-        "kea-waitfor": "^0.2.0",
+        "kea-waitfor": "^0.2.1",
         "kea-window-values": "^0.1.0",
         "md5": "^2.3.0",
         "posthog-js": "1.14.3",
@@ -158,6 +158,7 @@
         "html-webpack-plugin": "^4.5.2",
         "husky": "^4.3.0",
         "jest": "^26.6.3",
+        "kea-test-utils": "^0.1.2",
         "kea-typegen": "^1.2.2",
         "less": "^3.12.2",
         "less-loader": "^7.0.2",
